feat(union-ibc): test update client ko
