+++ cosmwasm/ibc-union/core/src/tests.rs
@@ -1,0 +1,12 @@
+mod client;
+
+use super::*;
+
+#[test]
+fn display() {
+    assert_eq!(
+        ContractErrorKind::ArithmeticOverflow,
+        ContractErrorKind::parse_from_error_message(&ContractError::ArithmeticOverflow.to_string())
+            .unwrap()
+    )
+}
