+++ cosmwasm/ibc-union/light-clients/cometbls/Cargo.toml
@@ -31,6 +31,7 @@ base64      = { workspace = true }
 hex-literal = { workspace = true }
 lazy_static = "1.4.0"
 serde_json  = { workspace = true }
+union-ibc = { workspace = true }
 
 
 [features]
