+++ cosmwasm/ibc-union/light-clients/cometbls/src/lib.rs
@@ -3,3 +3,6 @@ pub mod contract;
 pub mod errors;
 pub mod storage;
 pub mod zkp_verifier;
+
+#[cfg(test)]
+mod tests;
